refactor: Rename VPath -> VPlot
Move VPlot documents and code from raem to under @valos/plot library
