+++ packages/plot/VPlot-JSON.revdoc.test.js
@@ -6,9 +6,9 @@ const {
   },
 } = require("@valos/revdoc");
 
-const { conjoinVPath, disjoinVPathOutline, disjoinVPathString } = require(".");
+const { conjoinVPlot, disjoinVPlotOutline, disjoinVPlotString } = require(".");
 
-const title = "VPath JSON representations";
+const title = "VPlot JSON representations";
 const { itExpects, runTestDoc } = prepareTestDoc(title);
 
 module.exports = {
@@ -17,20 +17,21 @@ module.exports = {
     specStatus: "unofficial",
     editors: authors("iridian"),
     authors: authors("iridian"),
-    shortName: "vpathJSON",
+    shortName: "vplotJSON",
   },
   "chapter#abstract>0": {
-    "#0": [`VPath JSON is a JSON representation of VPath strings. It
-has two formats: generic 'VPath outlines' and their strict subset
-'sectioned VPaths'.
+    "#0": [
+`VPlot JSON is a JSON representation of VPlot strings. It
+has two formats: generic 'VPlot outlines' and their strict subset
+'sectioned VPlots'.
 
-Sectioned VPaths are the verbose, machine-processible canonical
-JSON 1-1 representation of VPath strings. Outlines are the more
+Sectioned VPlots are the verbose, machine-processible canonical
+JSON 1-1 representation of VPlot strings. Outlines are the more
 concise, writable and readable but non-canonical format: many different
-outlines can have the same VPath meaning. The normalization of an
-outline into its canonical, sectioned VPath is called disjoining.
+outlines can have the same VPlot meaning. The normalization of an
+outline into its canonical, sectioned VPlot is called disjoining.
 
-As sectioned VPaths are a subset of VPath outlines, a sectioned VPath
+As sectioned VPlots are a subset of VPlot outlines, a sectioned VPlot
 always disjoins into itself.
 `],
   },
@@ -45,17 +46,17 @@ always disjoins into itself.
 
   "chapter#introduction>2": {
     "#0": [
-`VPath JSON is a JSON representation of a VPath string.
-Different VPath JSON objects can express the same VPath which allows
+`VPlot JSON is a JSON representation of a VPlot string.
+Different VPlot JSON objects can express the same VPlot which allows
 the most suitable representation to be chosen based on use case.
 
-The canonical subset of VPath JSON (with 1-1 mapping to VPath strings)
-is called sectioned VPaths. These use a limited subset of JSON
+The canonical subset of VPlot JSON (with 1-1 mapping to VPlot strings)
+is called sectioned VPlots. These use a limited subset of JSON
 primitives and are designed to be easily machine-processible.
 
-Finally, the remaining, non-sectioned VPath JSON objects are called
-VPath outlines. Arbitrary JSON objects can be trivially escaped as
-VPath outline, allowing for generic JSON <-> VPath roundtripping.
+Finally, the remaining, non-sectioned VPlot JSON objects are called
+VPlot outlines. Arbitrary JSON objects can be trivially escaped as
+VPlot outline, allowing for generic JSON <-> VPlot roundtripping.
 Edit me - this is the first payload chapter. Abstract and SOTD are
 essential `, ref("ReSpec boilerplate",
     "https://github.com/w3c/respec/wiki/ReSpec-Editor's-Guide#essential-w3c-boilerplate"), `
@@ -68,23 +69,23 @@ ref("VDoc specification", "@valos/vdoc"), ` for reference documentation.`,
     ],
   },
 
-  "chapter#sectioned_vpath_json>2;Sectioned VPath JSON": {
+  "chapter#sectioned_vplot_json>2;Sectioned VPlot JSON": {
     "#0": [`
-Sectioned VPaths are the canonical JSON representation of a VPath.
+Sectioned VPlots are the canonical JSON representation of a VPlot.
 They use only a limited subset of JSON primitives: objects are not
 used; the literal 'null', the booleans 'true' and 'false', all string
 values and all safe integers (ie. whole numbers from -(2^53-1) to
 (2^53-1)) are used to represent their corresponding contextless vparam
 values directly.
 
-All remaining VPath elements are represented with two-entry arrays
+All remaining VPlot elements are represented with two-entry arrays
 as the titular 'sections'. The mandatory first entry contains a
 "@"-prefixed section type string. The optional second entry contains
 the section payload.
 
-The first two characters of the section type determine which vpath
+The first two characters of the section type determine which vplot
 element the section represents:`, [
-  [`'@@' denote a vpath element. The type has no other characters. The
+  [`'@@' denote a vplot element. The type has no other characters. The
 optional payload is an array of /two/ or more sections or values.`],
   [`'@$' denote a vparam element. The remaining type characters contain
 the vparam context-term. If there are none the vparam is contextless.
@@ -99,8 +100,8 @@ non-empty array of sections or values.`],
 
 `],
     "example#0": itExpects(
-        "forms a VPath string from VPath section of computing a random value between 2.5 and 10",
-() => conjoinVPath(
+        "forms a VPlot string from VPlot section of computing a random value between 2.5 and 10",
+() => conjoinVPlot(
     ["@@", [                // declare two-step path
       ["@!random"],         // first step: compute a random value
       ["@!min", [           // second step: compute the min of
@@ -116,30 +117,30 @@ non-empty array of sections or values.`],
 "@!random@!min$d.10$.@!max$d.2.5$.@@@@@@",
     ),
     "#1": [`
-During sectioning (be it from a vpath string or from an outline) all
+During sectioning (be it from a vplot string or from an outline) all
 redundant sections are elided or simplified:`,
 [
-  [`Single-step vpath sections are replaced with that step itself - a
-vpath represents a sequential dependency between its
-constitutient steps so a single-step vpath is redundant`],
+  [`Single-step vplot sections are replaced with that step itself - a
+vplot represents a sequential dependency between its
+constitutient steps so a single-step vplot is redundant`],
   [`VParam sections that represent null, true, false or any integer
 values are replaced with the direct JSON value`],
   [`Contextless vparam sections which contain a path or a verb section
 as the payload are replaced with the payload section itself.`],
 ],
 ``],
-    "example#1": itExpects("elides degenerate vpath elements during sectioning",
-() => disjoinVPathString("@!random@!min$d.10$.@!max$d.2.5$.@@@@@@"),
+    "example#1": itExpects("elides degenerate vplot elements during sectioning",
+() => disjoinVPlotString("@!random@!min$d.10$.@!max$d.2.5$.@@@@@@"),
 "toEqual",
 ["@@", [["@!random"], ["@!min", [10, ["@!max", [["@$d", "2.5"], ["@@"]]]]]]]
     ),
   },
-  "chapter#vpath_json_outline>2;VPath JSON outlines": {
+  "chapter#vplot_json_outline>2;VPlot JSON outlines": {
     "#0": [`
-VPath outlines are JSON structures which use convenience constructs.
+VPlot outlines are JSON structures which use convenience constructs.
 `],
     "example#0": itExpects("disjoins a configuration outline",
-() => disjoinVPathOutline([
+() => disjoinVPlotOutline([
   ["@$o.folder", "vault"],
   ["@$o.vlm", "@", "webpack"],
   ["@$o.folder", "build"],
