+++ packages/plot/_coerceOps.js
