+++ packages/plot/_validateTerminalOps.js
