+++ packages/raem/package.json
@@ -16,6 +16,7 @@
   "main": "index.js",
   "dependencies": {
     "@valos/tools": "^0.35.0-rc.35",
+    "@valos/plot": "^0.35.0-rc.35",
     "graphql": "^0.6.2",
     "immutable": "~3.7.4",
     "url-parse": "^1.4.1"
