+++ packages/sourcerer/tools/event-version-0.2/createVRID0Dot2.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { coerceAsVRID } from "~/raem/VPath";
+import { coerceAsVRID } from "~/plot";
 
 import { hashV240 } from "~/security/hash";
