+++ packages/web-spindle/schema-builder/_routesCommon.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { disjoinVPathOutline } from "~/raem/VPath";
+import { disjoinVPlotOutline } from "~/plot";
 import { patchWith } from "~/tools";
 
 import { ObjectSchema, StringType, IdValOSType } from "./types";
@@ -43,7 +43,7 @@ export function _finalizeRoute (route, userConfig, globalRules) {
     if (route.config.rules[ruleName] === undefined) return undefined;
   }
   for (const [key, rule] of Object.entries(route.config.rules)) {
-    route.config.rules[key] = disjoinVPathOutline(rule);
+    route.config.rules[key] = disjoinVPlotOutline(rule);
   }
   for (const ruleName of [].concat(
       route.config.requiredRules || [],
